requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
python-multipart
httpx
easyocr
redis
python-dotenv
